+++ AdminTools/AdminTools.csproj
@@ -59,30 +59,31 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Commands\Ahp\Ahp.cs" />
-    <Compile Include="Commands\Ball\Ball.cs" />
-    <Compile Include="Commands\BreakDoors\BreakDoors.cs" />
-    <Compile Include="Commands\DropItem\DropItem.cs" />
-    <Compile Include="Commands\DropSize\DropSize.cs" />
-    <Compile Include="Commands\Enums\Enums.cs" />
-    <Compile Include="Commands\Explode\Explode.cs" />
-    <Compile Include="Commands\FakeSync\FakeSpawnCommand.cs" />
-    <Compile Include="Commands\Ghost\Ghost.cs" />
-    <Compile Include="Commands\Grenade\Grenade.cs" />
-    <Compile Include="Commands\HintBroadcast\HintBroadcast.cs" />
-    <Compile Include="Commands\Jail\Jail.cs" />
-    <Compile Include="Commands\Position\Position.cs" />
-    <Compile Include="Commands\PryGates\PryGates.cs" />
-    <Compile Include="Commands\RandomTeleport\RandomTeleport.cs" />
-    <Compile Include="Commands\Regeneration\Regeneration.cs" />
-    <Compile Include="Commands\SpawnRagdoll\SpawnRagdoll.cs" />
-    <Compile Include="Commands\Scale\Scale.cs" />
-    <Compile Include="Commands\Size\Size.cs" />
-    <Compile Include="Commands\SpawnWorkbench\SpawnWorkbench.cs" />
+    <Compile Include="Commands\Ahp.cs" />
+    <Compile Include="Commands\Ball.cs" />
+    <Compile Include="Commands\BreakDoors.cs" />
+    <Compile Include="Commands\DropItem.cs" />
+    <Compile Include="Commands\DropSize.cs" />
+    <Compile Include="Commands\Enums.cs" />
+    <Compile Include="Commands\Explode.cs" />
+    <Compile Include="Commands\FakeSpawnCommand.cs" />
+    <Compile Include="Commands\Ghost.cs" />
+    <Compile Include="Commands\Grenade.cs" />
+    <Compile Include="Commands\HintBroadcast.cs" />
+    <Compile Include="Commands\InstantKill.cs" />
+    <Compile Include="Commands\Jail.cs" />
+    <Compile Include="Commands\Position.cs" />
+    <Compile Include="Commands\PryGates.cs" />
+    <Compile Include="Commands\RandomTeleport.cs" />
+    <Compile Include="Commands\Regeneration.cs" />
+    <Compile Include="Commands\Rocket.cs" />
+    <Compile Include="Commands\Scale.cs" />
+    <Compile Include="Commands\Size.cs" />
+    <Compile Include="Commands\SpawnRagdoll.cs" />
+    <Compile Include="Commands\SpawnWorkbench.cs" />
     <Compile Include="Commands\Tags\Hide.cs" />
     <Compile Include="Commands\Tags\Show.cs" />
     <Compile Include="Commands\Tags\Tags.cs" />
-    <Compile Include="Commands\InstantKill\InstantKill.cs" />
     <Compile Include="Commands\Inventory\Drop.cs" />
     <Compile Include="Commands\Inventory\Inventory.cs" />
     <Compile Include="Commands\Inventory\See.cs" />
@@ -90,9 +91,8 @@
     <Compile Include="Commands\Mute\ICom.cs" />
     <Compile Include="Commands\Mute\Mute.cs" />
     <Compile Include="Commands\Mute\RoundStart.cs" />
-    <Compile Include="Commands\Rocket\Rocket.cs" />
-    <Compile Include="Commands\TargetGhost\TargetGhost.cs" />
-    <Compile Include="Commands\TeleportX\TeleportX.cs" />
+    <Compile Include="Commands\TargetGhost.cs" />
+    <Compile Include="Commands\TeleportX.cs" />
     <Compile Include="Commands\Unmute\All.cs" />
     <Compile Include="Commands\Unmute\ICom.cs" />
     <Compile Include="Commands\Unmute\Unmute.cs" />
